requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.86.0
opencv-python==4.12.0.88
pillow==11.2.1
pluggy==1.6.0
pvcobra==2.0.5
pvleopard==2.0.5
pvorca==1.2.0
pvporcupine==3.0.5
pvrecorder==1.2.7
pvrhino==3.0.5
pvspeaker==1.0.5
pydantic==2.11.6
pydantic_core==2.33.2
pygame==2.6.1
requests==2.32.4
semver==3.0.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0